requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
requests
cloudinary
openai
sentence-transformers
scikit-learn
numpy
